Kernel: Start work on MSI and proper PCI-e support, ATA driver now makes use of a Semaphore, other minor fixes

+++ Kernel/include/math.h
@@ -4,5 +4,5 @@
 #define MIN(x, y) ((x > y) ? y : x)
 
 unsigned int rand();
-int floor(double num);
+int floor(int num);
 int abs(int num);
